Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/eventsource/eventsource-cross-origin-expected.txt

Blocked access to external URL http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=message
CONSOLE MESSAGE: EventSource cannot load http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=message due to access control checks.
Blocked access to external URL http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=redirect&location=/eventsource/resources/cors.py?run=message
CONSOLE MESSAGE: EventSource cannot load http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=redirect&location=/eventsource/resources/cors.py?run=message due to access control checks.
Blocked access to external URL http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=status-reconnect&status=200
CONSOLE MESSAGE: EventSource cannot load http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=status-reconnect&status=200 due to access control checks.
Blocked access to external URL http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=message&origin=http://example.org
CONSOLE MESSAGE: EventSource cannot load http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=message&origin=http://example.org due to access control checks.
Blocked access to external URL http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=message&origin=
CONSOLE MESSAGE: EventSource cannot load http://xn--lve-6lad.localhost:8800/eventsource/resources/cors.py?run=message&origin= due to access control checks.
Blocked access to external URL http://xn--lve-6lad.localhost:8800/eventsource/resources/message.py
CONSOLE MESSAGE: EventSource cannot load http://xn--lve-6lad.localhost:8800/eventsource/resources/message.py due to access control checks.

Harness Error (TIMEOUT), message = null

TIMEOUT EventSource: cross-origin basic use Test timed out
TIMEOUT EventSource: cross-origin redirect use Test timed out
TIMEOUT EventSource: cross-origin redirect use recon Test timed out
PASS EventSource: cross-origin allow-origin: http://example.org should fail
PASS EventSource: cross-origin allow-origin:'' should fail
PASS EventSource: cross-origin No allow-origin should fail
